Move RecentCouchChangesView to views/system.py
