+++ corehq/apps/hqadmin/urls.py
@@ -4,9 +4,7 @@ from django.conf.urls import include, url
 from corehq.apps.domain.decorators import require_superuser
 from corehq.apps.domain.utils import new_domain_re
 from corehq.apps.hqadmin.views.data import (
-    RecentCouchChangesView,
     doc_in_es,
-    download_recent_changes,
     raw_couch,
     raw_doc,
 )
@@ -25,9 +23,11 @@ from corehq.apps.hqadmin.views.reports import (
     top_five_projects_by_country,
 )
 from corehq.apps.hqadmin.views.system import (
+    RecentCouchChangesView,
     SystemInfoView,
     branches_on_staging,
     check_services,
+    download_recent_changes,
     pillow_operation_api,
     system_ajax,
 )
