+++ corehq/apps/hqadmin/views/__init__.py
@@ -1,9 +1,7 @@
 from __future__ import absolute_import
 
 from corehq.apps.hqadmin.views.data import (
-    RecentCouchChangesView,
     doc_in_es,
-    download_recent_changes,
     get_db_from_db_name,
     raw_couch,
     raw_doc,
@@ -23,9 +21,11 @@ from corehq.apps.hqadmin.views.reports import (
     top_five_projects_by_country,
 )
 from corehq.apps.hqadmin.views.system import (
+    RecentCouchChangesView,
     SystemInfoView,
     branches_on_staging,
     check_services,
+    download_recent_changes,
     get_rabbitmq_management_url,
     pillow_operation_api,
     system_ajax,
