+++ corehq/apps/hqadmin/views/data.py
@@ -248,47 +248,3 @@ def raw_doc(request):
     context['all_databases'] = ['commcarehq'] + other_couch_dbs + list(_SQL_DBS)
     context['use_code_mirror'] = request.GET.get('code_mirror', 'true').lower() == 'true'
     return render(request, "hqadmin/raw_couch.html", context)
-
-
-class RecentCouchChangesView(BaseAdminSectionView):
-    urlname = 'view_recent_changes'
-    template_name = 'hqadmin/couch_changes.html'
-    page_title = ugettext_lazy("Recent Couch Changes")
-
-    @use_nvd3_v3
-    @use_datatables
-    @use_jquery_ui
-    @method_decorator(require_superuser_or_contractor)
-    def dispatch(self, *args, **kwargs):
-        return super(RecentCouchChangesView, self).dispatch(*args, **kwargs)
-
-    @property
-    def page_context(self):
-        count = int(self.request.GET.get('changes', 1000))
-        changes = list(get_recent_changes(get_db(), count))
-        domain_counts = defaultdict(lambda: 0)
-        doc_type_counts = defaultdict(lambda: 0)
-        for change in changes:
-            domain_counts[change['domain']] += 1
-            doc_type_counts[change['doc_type']] += 1
-
-        def _to_chart_data(data_dict):
-            return [
-                {'label': l, 'value': v} for l, v in sorted(list(data_dict.items()), key=lambda tup: tup[1], reverse=True)
-            ][:20]
-
-        return {
-            'count': count,
-            'recent_changes': changes,
-            'domain_data': {'key': 'domains', 'values': _to_chart_data(domain_counts)},
-            'doc_type_data': {'key': 'doc types', 'values': _to_chart_data(doc_type_counts)},
-        }
-
-
-@require_superuser_or_contractor
-def download_recent_changes(request):
-    count = int(request.GET.get('changes', 10000))
-    resp = HttpResponse(content_type='text/csv')
-    resp['Content-Disposition'] = 'attachment; filename="recent_changes.csv"'
-    download_changes(get_db(), count, resp)
-    return resp
